- cooldown -> entity - position separated from box

+++ DarkDefenders.Domain.Model/EntityProperties/CreatureProperties.cs
@@ -4,10 +4,10 @@ namespace DarkDefenders.Domain.Model.EntityProperties
 {
     public class CreatureProperties: SlowValueObject
     {
-        public float MovementForce { get; private set; }
-        public float JumpMomentum { get; private set; }
+        public double MovementForce { get; private set; }
+        public double JumpMomentum { get; private set; }
 
-        public CreatureProperties(float movementForce, float jumpMomentum)
+        public CreatureProperties(double movementForce, double jumpMomentum)
         {
             MovementForce = movementForce;
             JumpMomentum = jumpMomentum;
